requirement.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
numpy==2.0.2
opencv-python==4.11.0.86
pyserial==3.5
requests==2.32.3
urllib3==2.4.0
